Fixes #64
Compilation now leverages the ast to enable direct inclusion of the complied expression for an identifier element that has its token over-ridden creating a new first level rule provided that the referenced identifier is not itself an inline created token (i.e. is a top level rule). More work would be required to achieve that (parsing all elements, and identifying those with token overrides and compiling them, with the associated recursion challenges where they would have to remain bound references to the identifier as sequences of a single item)