Added requirement for grammar declaration
Added the requirement for the name of the grammar to be declared as the first declaration in the STLR file. This will enable future extensions supporting name spacing.

This does break existing STLR grammars but is easy to resolve and has a clear error message.

Updated tests to deal with this change