Fixed STLR generation issue
Fixed the issue with STLR that meant the grammar name was not correctly generated.

There are bugs with wrapping identifiers in modifiers that should be resolved by the new rules engine (that is, resolved such that they are significantly more simple) so I am not going to fix them at this point.

Also extended the various stlrc commands to consume the grammar name