Fixed a defect with negated rules
Negated rules were not correctly evaluated. The fix actually simplifies the evaluate function and the issue was highlighted by the unit tests and the enforced precidence rules. So I'm feeling much more confident and with precidence rules not encapsulated in a single class if changes need to be made there is a single point to do that at.